#20 Last Update

+++ database/migrations/2024_04_20_140823_add_borrowed_count_to_books_table.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->dropColumn('borrowed_count');
+            if (Schema::hasColumn('books', 'borrowed_count')) {
+                $table->dropColumn('borrowed_count');
+            }
         });
     }
 };
